Refactor project structure.
